# docs/openapi/phase2-openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
...
      "RoleAssignmentRequest": {
        "properties": {
          "branch_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "company_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "membership_id": {
            "format": "uuid",
            "title": "Membership Id",
            "type": "string"
          },
          "role_id": {
            "format": "uuid",
            "title": "Role Id",
            "type": "string"
          }
        },
        "required": [
          "membership_id",
          "role_id"
        ],
        "title": "RoleAssignmentRequest",
        "type": "object"
      },
      "RoleAssignmentResponse": {
        "properties": {
          "assigned_by_user_id": {
            "format": "uuid",
            "title": "Assigned By User Id",
            "type": "string"
          },
          "branch_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "company_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "membership_id": {
            "format": "uuid",
            "title": "Membership Id",
            "type": "string"
          },
          "role_id": {
            "format": "uuid",
            "title": "Role Id",
            "type": "string"
          },
          "tenant_id": {
            "format": "uuid",
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "membership_id",
          "role_id",
          "company_id",
          "branch_id",
          "assigned_by_user_id"
        ],
        "title": "RoleAssignmentResponse",
        "type": "object"
      },
      "RoleCreateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "permission_keys": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Permission Keys",
            "type": "array"
          },
          "scope": {
            "pattern": "^(tenant|company|branch)$",
            "title": "Scope",
            "type": "string"
          },
          "slug": {
            "maxLength": 100,
            "minLength": 2,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug",
          "scope",
          "permission_keys"
        ],
        "title": "RoleCreateRequest",
        "type": "object"
      },
      "RoleResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_editable": {
            "title": "Is Editable",
            "type": "boolean"
          },
          "is_system": {
            "title": "Is System",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tenant_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "name",
          "slug",
          "scope",
          "is_system",
          "is_editable",
          "version",
          "permissions"
        ],
        "title": "RoleResponse",
        "type": "object"
      },
      "RoleUpdateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 400,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "permission_keys": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Permission Keys"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "RoleUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Pharma Intelligence SaaS",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/roles": {
      "get": {
        "operationId": "get_roles_api_v1_roles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleResponse"
                  },
                  "title": "Response Get Roles Api V1 Roles Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Roles",
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "post_role_api_v1_roles_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Role",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/assignments": {
      "get": {
        "operationId": "get_role_assignments_api_v1_roles_assignments_get",
        "parameters": [
          {
            "in": "query",
            "name": "membership_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Membership Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleAssignmentResponse"
                  },
                  "title": "Response Get Role Assignments Api V1 Roles Assignments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Role Assignments",
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "post_role_assignment_api_v1_roles_assignments_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleAssignmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Role Assignment",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/assignments/{assignment_id}": {
      "delete": {
        "operationId": "delete_role_assignment_api_v1_roles_assignments__assignment_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Role Assignment",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{role_id}": {
      "delete": {
        "operationId": "remove_role_api_v1_roles__role_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Role",
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "patch_role_api_v1_roles__role_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Role",
        "tags": [
          "roles"
        ]
      }
    },
...
  }
}